requirements.txt
kivy==2.2.1
kivymd==1.1.1
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0
numpy==1.24.3
opencv-python==4.8.0.74
pyzbar==0.1.9
